# getPids - don't recursively traverse every dir in /proc