--- a/pkg/util/procfs/procfs_linux_test.go
+++ b/pkg/util/procfs/procfs_linux_test.go
@@ -23,6 +23,7 @@ import (
 	"os"
 	"os/signal"
 	"path/filepath"
+	"regexp"
 	"runtime"
 	"syscall"
 	"testing"
@@ -93,5 +94,23 @@ func TestPKill(t *testing.T) {
 		}
 	case <-time.After(1 * time.Second):
 		t.Fatalf("timeout waiting for %v", sig)
+	}
+}
+
+func BenchmarkGetPids(b *testing.B) {
+	if runtime.GOOS == "darwin" || runtime.GOOS == "windows" {
+		b.Skipf("not supported on GOOS=%s", runtime.GOOS)
+	}
+
+	re, err := regexp.Compile("(^|/)" + filepath.Base(os.Args[0]) + "$")
+	assert.Empty(b, err)
+
+	for i := 0; i < b.N; i++ {
+		pids := getPids(re)
+
+		b.StopTimer()
+		assert.NotZero(b, pids)
+		assert.Contains(b, pids, os.Getpid())
+		b.StartTimer()
 	}
 }