**What this PR does / why we need it**:

`filepath.Walk` recursively traverses every dir, which is not what is
needed for `getPids`.
Instead only read the list of dirs in the top level of `/proc`.

```
benchmark              old ns/op     new ns/op     delta
BenchmarkGetPids-4     868684        195522        -77.49%
```

```release-note
Speedup process lookup in /proc
```