@cpuguy83 2018-07-19
Updated, also optimized in case of many entries in `/proc` so that we don't allocate a huge slice.