@fcrisciani 2018-10-29
still valid 
/remove-lifecycle stale