@thockin 2018-10-29
Would it be simpler to make a more surgical change?  Something like:

```
    filepath.Walk("/proc", func(path string, info os.FileInfo, err error) error {
        if err != nil {
            // We should continue processing other directories/files
            return nil 
        }
        if path == "/proc" {
            // Skip the top-level dir, but recurse
            return nil 
        }
        fmt.Printf("path %v\n", path)
        if !info.IsDir() {
            // Skip files
            return nil 
        }
        // If the directory is not a number (i.e. not a PID), skip it
        base := filepath.Base(path)
        if _, err := strconv.Atoi(base); err != nil {
            fmt.Printf("skipping %v\n", path)
            return filepath.SkipDir
        }
        fmt.Printf("found pid %v => %v/cmdline\n", base, path)
        // ...process it...
        return filepath.SkipDir // done with this dir, don't recurse further
    })  
```