@thockin 2018-10-29
Did some quick tests, this is better and complete